pyproject.toml
[tool.poetry]
name = "content-creating-team"
version = "0.1.0"
description = "Simple AI content planner"
authors = ["Prasad A Parit"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"  # Updated to allow Python 3.12.x
crewai = "*"
unstructured = "==0.10.25"
pyowm = "3.3.0"
python-dotenv = "1.0.0"
python-decouple = "^3.8"
langchain-openai = "0.0.5"
streamlit = "^1.41.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
